handle database down migrations

+++ database/migrations/2019_12_05_162050_create_jobs_table.php
@@ -4,15 +4,9 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreateJobsTable extends Migration
-{
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
-    public function up()
-    {
+class CreateJobsTable extends Migration {
+
+    public function up() {
         Schema::create('jobs', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('queue')->index();
@@ -24,13 +18,9 @@ class CreateJobsTable extends Migration
         });
     }
 
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
-    public function down()
-    {
-        Schema::dropIfExists('jobs');
+    public function down() {
+        if (Schema::hasTable('jobs')) {
+            Schema::dropIfExists('jobs');
+        }
     }
 }
